fix: resolve six 500-class bugs surfaced by the Phase 1 test suite
- routers/admin: catch ValueError on webhook get/delete for unknown
  IDs (500 -> 404); rewire patch_webhook to admin_webhook_patch (it
  previously called admin_webhook_get with a payload kwarg, raising
  TypeError on every call)
- routers/drafts: catch ValueError when sending an unknown draft ID
  (500 -> 404)
- auth: a valid JWT whose subject no longer exists now returns 401
  instead of propagating the backend's ValueError as a 500
- validators: catch ValueError instead of pydantic.ValidationError so
  unparseable JSON bodies 422 like invalid ones (JSONDecodeError also
  subclasses ValueError) across all body-validated endpoints
- mail-server: declare python-multipart (required by the OAuth2 login
  form on POST /auth/token; previously only present transitively)

The strict xfails pinning these bugs are flipped into permanent
regression tests; PATCH /admin/webhooks/{id} now surfaces as a clean
NotImplementedError stub tracked in test_stubs.py.
