+++ src/mail/server/pyproject.toml
@@ -15,6 +15,7 @@ dependencies = [
     "pydantic>=2.11.7",
     "pyjwt>=2.10.1",
     "python-dotenv>=1.1.1",
+    "python-multipart>=0.0.20",
     "uvicorn>=0.35.0",
 ]
